Refactor: 리뷰 반영 + model 패키지 구조 세분화
model 내부의 패키지를 ladder, user prize로 세분화
생성자 private로 변경
테스트 코드 for loop -> Stream API 변경
