+++ src/main/java/model/MoveCondition.java
@@ -1,5 +1,7 @@
 package model;
 
+import model.ladder.HorizontalLine;
+
 @FunctionalInterface
 public interface MoveCondition {
     boolean canMove(int index, HorizontalLine line);
