+++ src/main/java/model/Prize/Prize.java
@@ -1,4 +1,4 @@
-package model;
+package model.Prize;
 
 public class Prize {
