+++ src/main/java/model/ladder/Link.java
@@ -1,4 +1,4 @@
-package model;
+package model.ladder;
 
 public enum Link {
     LINKED,
