+++ src/main/java/model/ladder/LinkGenerator.java
@@ -1,4 +1,4 @@
-package model;
+package model.ladder;
 
 public interface LinkGenerator {
     Link generate();
