+++ src/main/java/model/user/Users.java
@@ -1,4 +1,4 @@
-package model;
+package model.user;
 
 import java.util.Collections;
 import java.util.HashSet;
