+++ src/test/java/model/LadderGameResultTest.java
@@ -1,5 +1,10 @@
 package model;
 
+import model.Prize.Prize;
+import model.Prize.Prizes;
+import model.ladder.Ladder;
+import model.user.User;
+import model.user.Users;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
