+++ src/test/java/model/LinkTest.java
@@ -1,5 +1,6 @@
 package model;
 
+import model.ladder.Link;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.*;
