+++ src/test/java/model/LinkedGenerator.java
@@ -1,5 +1,8 @@
 package model;
 
+import model.ladder.Link;
+import model.ladder.LinkGenerator;
+
 public class LinkedGenerator implements LinkGenerator {
     @Override
     public Link generate() {
