+++ src/test/java/model/MovingDirectionTest.java
@@ -1,5 +1,7 @@
 package model;
 
+import model.ladder.HorizontalLine;
+import model.ladder.LineGenerator;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.*;
