+++ src/test/java/model/UnLinkedGenerator.java
@@ -1,6 +1,9 @@
 package model;
 
-public class UnLinkedGenerator implements LinkGenerator{
+import model.ladder.Link;
+import model.ladder.LinkGenerator;
+
+public class UnLinkedGenerator implements LinkGenerator {
     @Override
     public Link generate() {
         return Link.from(false);
